# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Consulta de usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "UnAuthorize"
                ],
                "summary": "Autenticación de usuarios",
                "operationId": "f74109d26aab35daab2d8c617c97c3fa",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "george.bluth@reqres.in"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Authorize"
                ],
                "summary": "Mostrar todos los usuarios.",
                "operationId": "60e6ec72fbb2dd5f8977b3623a5bd2f6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de pagina actual",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Authorize"
                ],
                "summary": "Creación de usuarios.",
                "operationId": "23b7ba485f2f33276fee0b673f729a51",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "first_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "tyesid@hotmail.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Yecid"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Authorize"
                ],
                "summary": "Mostrar detalle del usuario",
                "operationId": "1b0033ebfd92ac053df80ca3523c618d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Authorize"
                ],
                "summary": "Actualizar usuarios",
                "operationId": "30d7b826acc88364c63d5c76e52c5ca8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de usuario",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "first_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "tyesid@hotmail.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Yecid Update"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Tovar Update"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Inicie sesión con correo electrónico y contraseña para obtener el token de autenticación",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}